+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+
+## Unreleased
+
+ -  Check container availability before the configuration
+
 ## 0.0.1 - 2017-06-26
 
  - Implementation of core features